\name{allCounts}
\docType{data}
\alias{allCounts}
\alias{Counts}
\title{Region (exons and junctions) counts}
\description{
Region (exons and junctions) counts from BAM samples in package \code{\link[RNAseqData.HNRNPC.bam.chr14]{RNAseqData.HNRNPC.bam.chr14}}
}
\usage{data(Counts)}
\format{A matrix named with 881654 rows (exons and
  junctions) and 8 columns samples}
\keyword{datasets}
\keyword{RNA-seq}
