\name{txdb}
\docType{data}
\alias{txdb}
\alias{TxDb}
\title{Reshaped TranscriptDb annotation object}
\description{
Annotation object for human chromosome 14 computed from package \code{\link[TxDb.Hsapiens.UCSC.hg19.knownGene]{TxDb.Hsapiens.UCSC.hg19.knownGene}}
}
\usage{data(TxDb)}
\value{A \code{\linkS4class{GRangesList}} object of length 634}
\keyword{datasets}
\keyword{RNA-seq}
